% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{descendancy_chart}
\alias{descendancy_chart}
\title{Generate a descendancy chart}
\usage{
descendancy_chart(gedcom, indi_xref = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}
}
\value{
A descendancy chart showing the descendants of the individual.
}
\description{
Generate a descendancy chart
}
