% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{family_group_chart}
\alias{family_group_chart}
\title{Generate a family group chart}
\usage{
family_group_chart(gedcom, family)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{family}{An xref identifying a Family group record.}
}
\value{
A chart showing a family group.
}
\description{
Generate a family group chart
}
