% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{node_label}
\alias{node_label}
\title{Create node labels for pedigree and descendancy charts}
\usage{
node_label(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the Individual or Family Group record.}
}
\value{
A string describing a node to use in the DiagrammeR::mermaid() function.
}
\description{
This function creates brief summaries of individuals and relationships formatted
for the DiagrammeR::mermaid() function.
}
