% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{node_style}
\alias{node_style}
\title{Create node styles for pedigree and descendancy charts}
\usage{
node_style(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the Individual or Family Group record.}
}
\value{
A string describing a node style to use in the DiagrammeR::mermaid() function.
}
\description{
This function creates node styles formatted for the DiagrammeR::mermaid() function.
}
