% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pedigree_chart}
\alias{pedigree_chart}
\title{Generate a pedigree chart}
\usage{
pedigree_chart(gedcom, individual = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{individual}{An xref or character string identifying an Individual record.}
}
\value{
A pedigree chart showing the ancestors of the individual.
}
\description{
Generate a pedigree chart
}
