% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevis.R
\name{timevis_famg}
\alias{timevis_famg}
\title{Construct a fact timeline for a family group}
\usage{
timevis_famg(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the family group to be plotted.}
}
\value{
A timevis object.
}
\description{
Construct a fact timeline for a family group
}
\details{
Only facts with associated dates are plotted.
}
