% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevis.R
\name{timevis_indi}
\alias{timevis_indi}
\title{Construct a fact timeline for an individual}
\usage{
timevis_indi(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the individual.}
}
\value{
A timevis object.
}
\description{
Construct a fact timeline for an individual
}
\details{
Only facts with associated dates are plotted.
}
