% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visged-package.R
\docType{package}
\name{visged-package}
\alias{visged}
\alias{visged-package}
\title{visged: Visualise GEDCOM Files}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Produce a variety of visualisations for family tree GEDCOM files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jl5000/visged}
}

}
\author{
\strong{Maintainer}: Jamie Lendrum \email{jalendrum@gmail.com}

}
\keyword{internal}
