% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{descendancy_chart}
\alias{descendancy_chart}
\title{Generate a descendancy chart}
\usage{
descendancy_chart(gedcom, individual = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{individual}{An xref or character string identifying an Individual record.}
}
\value{
A descendancy chart showing the descendants of the individual.
}
\description{
Generate a descendancy chart
}
