% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{pedigree_chart}
\alias{pedigree_chart}
\title{Generate a pedigree chart}
\usage{
pedigree_chart(gedcom, indi_xref = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}
}
\value{
A pedigree chart showing the ancestors of the individual.
}
\description{
Generate a pedigree chart
}
